% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_versions.R
\name{vmx_data_version}
\alias{vmx_data_version}
\title{Fetch one data version}
\usage{
vmx_data_version(id, client = vmx_client())
}
\arguments{
\item{id}{A data-version id (\code{dv_...}) or \code{vmx_data_version}.}

\item{client}{A \code{vmx_client}.}
}
\value{
A \code{vmx_data_version}.
}
\description{
Fetch one data version
}
