% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_versions.R
\name{vmx_data_version_unarchive}
\alias{vmx_data_version_unarchive}
\title{Unarchive a data version}
\usage{
vmx_data_version_unarchive(dv, client = vmx_client())
}
\arguments{
\item{dv}{A data-version id or \code{vmx_data_version}.}

\item{client}{A \code{vmx_client}.}
}
\description{
Unarchive a data version
}
