% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{vmx_dataset}
\alias{vmx_dataset}
\title{Fetch one dataset}
\usage{
vmx_dataset(id, client = vmx_client())
}
\arguments{
\item{id}{A dataset id (\code{ds_...}) or \code{vmx_dataset}.}

\item{client}{A \code{vmx_client}.}
}
\value{
A \code{vmx_dataset}.
}
\description{
Fetch one dataset
}
