% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{vmx_dataset_download}
\alias{vmx_dataset_download}
\title{Download a dataset's files}
\usage{
vmx_dataset_download(dataset, dest = ".", client = vmx_client())
}
\arguments{
\item{dataset}{A dataset id or \code{vmx_dataset}.}

\item{dest}{Destination directory.}

\item{client}{A \code{vmx_client}.}
}
\value{
Character vector of written file paths.
}
\description{
Download a dataset's files
}
