% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{vmx_dataset_files}
\alias{vmx_dataset_files}
\title{List the files in a dataset}
\usage{
vmx_dataset_files(dataset, client = vmx_client())
}
\arguments{
\item{dataset}{A dataset id or \code{vmx_dataset}.}

\item{client}{A \code{vmx_client}.}
}
\value{
A tibble containing all files in the dataset.
}
\description{
List the files in a dataset
}
