% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{vmx_dataset_tags}
\alias{vmx_dataset_tags}
\title{List the tags on a dataset}
\usage{
vmx_dataset_tags(dataset, client = vmx_client())
}
\arguments{
\item{dataset}{A dataset id or \code{vmx_dataset}.}

\item{client}{A \code{vmx_client}.}
}
\value{
A tibble.
}
\description{
List the tags on a dataset
}
