% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{vmx_dosing_input}
\alias{vmx_dosing_input}
\title{Create a dosing input for a fit}
\usage{
vmx_dosing_input(fit, dosing_text, scenario_name, client = vmx_client())
}
\arguments{
\item{fit}{A fit id or \code{vmx_model_fit}.}

\item{dosing_text}{Dosing regimen text.}

\item{scenario_name}{Scenario label.}

\item{client}{A \code{vmx_client}.}
}
\value{
A \code{vmx_dosing_input}.
}
\description{
Create a dosing input for a fit
}
