% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{vmx_dosing_input_status}
\alias{vmx_dosing_input_status}
\title{Dosing-input status}
\usage{
vmx_dosing_input_status(dosing_input, client = vmx_client())
}
\arguments{
\item{dosing_input}{A dosing-input id or \code{vmx_dosing_input}.}

\item{client}{A \code{vmx_client}.}
}
\value{
A \code{vmx_dosing_input}.
}
\description{
Dosing-input status
}
