% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resources.R
\name{vmx_has_next_page}
\alias{vmx_has_next_page}
\title{Test whether an API page has a following page}
\usage{
vmx_has_next_page(x)
}
\arguments{
\item{x}{A tibble returned by a vmxr collection function.}
}
\value{
A single logical value.
}
\description{
Test whether an API page has a following page
}
