% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{vmx_model_build_artifacts}
\alias{vmx_model_build_artifacts}
\title{Download build-run artifacts}
\usage{
vmx_model_build_artifacts(run, dest = ".", client = vmx_client())
}
\arguments{
\item{run}{A build-run id or object.}

\item{dest}{Destination directory.}

\item{client}{A \code{vmx_client}.}
}
\value{
Not yet implemented.
}
\description{
Download build-run artifacts
}
