% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{vmx_model_build_cancel}
\alias{vmx_model_build_cancel}
\title{Cancel a build run}
\usage{
vmx_model_build_cancel(run, client = vmx_client())
}
\arguments{
\item{run}{A build-run id or object.}

\item{client}{A \code{vmx_client}.}
}
\value{
A \code{vmx_model_build_run}.
}
\description{
Cancel a build run
}
