% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{vmx_model_build_events}
\alias{vmx_model_build_events}
\title{Build-run events (SSE stream)}
\usage{
vmx_model_build_events(run, client = vmx_client())
}
\arguments{
\item{run}{A build-run id or object.}

\item{client}{A \code{vmx_client}.}
}
\value{
Not yet implemented.
}
\description{
Build-run events (SSE stream)
}
