% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{vmx_model_build_logs}
\alias{vmx_model_build_logs}
\title{Build-run logs}
\usage{
vmx_model_build_logs(run, client = vmx_client())
}
\arguments{
\item{run}{A build-run id or object.}

\item{client}{A \code{vmx_client}.}
}
\description{
Build-run logs
}
