% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{vmx_model_build_status}
\alias{vmx_model_build_status}
\title{Build-run status}
\usage{
vmx_model_build_status(run, client = vmx_client())
}
\arguments{
\item{run}{A build-run id (\code{run_...}) or object.}

\item{client}{A \code{vmx_client}.}
}
\value{
A \code{vmx_model_build_run}.
}
\description{
Build-run status
}
