% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{vmx_model_describe}
\alias{vmx_model_describe}
\title{Describe a model}
\usage{
vmx_model_describe(model_name, client = vmx_client())
}
\arguments{
\item{model_name}{Model name.}

\item{client}{A \code{vmx_client}.}
}
\value{
A list / S3 object.
}
\description{
Describe a model
}
