% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{vmx_model_fit_postprocessor_status}
\alias{vmx_model_fit_postprocessor_status}
\title{Model-fit postprocessor status}
\usage{
vmx_model_fit_postprocessor_status(fit, client = vmx_client())
}
\arguments{
\item{fit}{A fit id or \code{vmx_model_fit}.}

\item{client}{A \code{vmx_client}.}
}
\value{
A list with postprocessor status.
}
\description{
Model-fit postprocessor status
}
