% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{vmx_model_fits}
\alias{vmx_model_fits}
\title{List model fits}
\usage{
vmx_model_fits(
  run = NULL,
  data_version = NULL,
  model_type = NULL,
  marker_name = NULL,
  status = NULL,
  client = vmx_client(),
  treatment = NULL,
  study = NULL,
  source_pk_model_fit = NULL
)
}
\arguments{
\item{run}{Optional model-build run filter.}

\item{data_version}{Optional data-version filter.}

\item{model_type}{Optional model-type filter.}

\item{marker_name}{Optional marker-name filter.}

\item{status}{Optional model-fit status filter.}

\item{client}{A \code{vmx_client}.}

\item{treatment}{Optional treatment filter.}

\item{study}{Optional study filter.}

\item{source_pk_model_fit}{Optional source PK model-fit filter.}
}
\value{
A tibble containing all matching model fits.
}
\description{
List model fits
}
