% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nca.R
\name{vmx_nca_get}
\alias{vmx_nca_get}
\title{Fetch one NCA analysis}
\usage{
vmx_nca_get(id, client = vmx_client())
}
\arguments{
\item{id}{An NCA id (\code{nca_...}) or \code{vmx_nca_analysis}.}

\item{client}{A \code{vmx_client}.}
}
\value{
A \code{vmx_nca_analysis}.
}
\description{
Fetch one NCA analysis
}
