% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_versions.R
\name{vmx_pk}
\alias{vmx_pk}
\title{PK observations table}
\usage{
vmx_pk(dv, time_basis = NULL, client = vmx_client())
}
\arguments{
\item{dv}{A data-version id or \code{vmx_data_version}.}

\item{time_basis}{Optional time basis; see \code{\link[=vmx_data_version_table]{vmx_data_version_table()}}.}

\item{client}{A \code{vmx_client}.}
}
\value{
A tibble.
}
\description{
PK observations table
}
