% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.R
\name{vmx_prep_questions}
\alias{vmx_prep_questions}
\title{Questions raised by prep (when awaiting input)}
\usage{
vmx_prep_questions(dataset, client = vmx_client())
}
\arguments{
\item{dataset}{A dataset id or \code{vmx_dataset}.}

\item{client}{A \code{vmx_client}.}
}
\value{
A tibble of pending questions.
}
\description{
Questions raised by prep (when awaiting input)
}
