% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{vmx_sim_cancel}
\alias{vmx_sim_cancel}
\title{Cancel a simulation job}
\usage{
vmx_sim_cancel(job, client = vmx_client())
}
\arguments{
\item{job}{A job id or \code{vmx_simulation_job}.}

\item{client}{A \code{vmx_client}.}
}
\value{
A \code{vmx_simulation_job}.
}
\description{
Cancel a simulation job
}
