% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{vmx_sim_existing_subject}
\alias{vmx_sim_existing_subject}
\title{Simulate an existing subject}
\usage{
vmx_sim_existing_subject(fit, spec, wait = FALSE, client = vmx_client())
}
\arguments{
\item{fit}{A fit id or \code{vmx_model_fit}.}

\item{spec}{Simulation spec.}

\item{wait}{If \code{TRUE}, block until the job settles.}

\item{client}{A \code{vmx_client}.}
}
\value{
A \code{vmx_simulation_job}.
}
\description{
Simulate an existing subject
}
