% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{vmx_sim_result}
\alias{vmx_sim_result}
\title{Simulation result table}
\usage{
vmx_sim_result(job, client = vmx_client())
}
\arguments{
\item{job}{A job id or \code{vmx_simulation_job}.}

\item{client}{A \code{vmx_client}.}
}
\value{
A tibble.
}
\description{
Simulation result table
}
