% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{vmx_sim_status}
\alias{vmx_sim_status}
\title{Simulation job status}
\usage{
vmx_sim_status(job, client = vmx_client())
}
\arguments{
\item{job}{A job id or \code{vmx_simulation_job}.}

\item{client}{A \code{vmx_client}.}
}
\value{
A \code{vmx_simulation_job}.
}
\description{
Simulation job status
}
