% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studies.R
\name{vmx_study}
\alias{vmx_study}
\title{Fetch one study}
\usage{
vmx_study(id, client = vmx_client())
}
\arguments{
\item{id}{A study id (\code{std_...}) or \code{vmx_study} object.}

\item{client}{A \code{vmx_client}.}
}
\value{
A \code{vmx_study}.
}
\description{
Fetch one study
}
