% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studies.R
\name{vmx_study_update}
\alias{vmx_study_update}
\title{Update a study}
\usage{
vmx_study_update(id, ..., client = vmx_client())
}
\arguments{
\item{id}{A study id or \code{vmx_study}.}

\item{...}{Fields to update.}

\item{client}{A \code{vmx_client}.}
}
\value{
A \code{vmx_study}.
}
\description{
Update a study
}
