% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treatments.R
\name{vmx_treatment}
\alias{vmx_treatment}
\title{Fetch one treatment}
\usage{
vmx_treatment(id, client = vmx_client())
}
\arguments{
\item{id}{A treatment id (\code{tmt_...}) or \code{vmx_treatment} object.}

\item{client}{A \code{vmx_client}.}
}
\value{
A \code{vmx_treatment}.
}
\description{
Fetch one treatment
}
