% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treatments.R
\name{vmx_treatment_create}
\alias{vmx_treatment_create}
\title{Create a treatment}
\usage{
vmx_treatment_create(
  name,
  indication = NULL,
  description = NULL,
  client = vmx_client()
)
}
\arguments{
\item{name}{Treatment name.}

\item{indication}{Optional indication.}

\item{description}{Optional free-text description.}

\item{client}{A \code{vmx_client}.}
}
\value{
A \code{vmx_treatment}.
}
\description{
Create a treatment
}
