% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{vmx_upload_unignore}
\alias{vmx_upload_unignore}
\title{Reverse a prior upload ignore}
\usage{
vmx_upload_unignore(dataset, upload, client = vmx_client())
}
\arguments{
\item{dataset}{A dataset id or \code{vmx_dataset}.}

\item{upload}{The upload id (\code{upl_...}) to ignore.}

\item{client}{A \code{vmx_client}.}
}
\value{
The updated \code{vmx_prep_status}.
}
\description{
\code{POST /datasets/{ds_id}/unignore-upload}.
}
