% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_versions.R
\name{vmx_data_version_create}
\alias{vmx_data_version_create}
\title{Create a data version}
\usage{
vmx_data_version_create(
  dataset,
  uploads,
  prior_config = NULL,
  client = vmx_client()
)
}
\arguments{
\item{dataset}{A dataset id or \code{vmx_dataset}.}

\item{uploads}{Uploads to include.}

\item{prior_config}{Optional prior config for warm-start.}

\item{client}{A \code{vmx_client}.}
}
\value{
A \code{vmx_data_version}.
}
\description{
Create a data version
}
