% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_versions.R
\name{vmx_data_version_export}
\alias{vmx_data_version_export}
\title{Export a data version}
\usage{
vmx_data_version_export(dv, dest = NULL, client = vmx_client())
}
\arguments{
\item{dv}{A data-version id or \code{vmx_data_version}.}

\item{dest}{Optional destination; when \code{NULL}, returns the data.}

\item{client}{A \code{vmx_client}.}
}
\description{
Export a data version
}
