% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_versions.R
\name{vmx_data_versions}
\alias{vmx_data_versions}
\title{List data versions}
\usage{
vmx_data_versions(
  treatment = NULL,
  study = NULL,
  include_archived = FALSE,
  eligible_for_modeling = NULL,
  client = vmx_client()
)
}
\arguments{
\item{treatment}{Optional treatment filter.}

\item{study}{Optional study filter.}

\item{include_archived}{Include archived versions.}

\item{eligible_for_modeling}{Optional modeling-eligibility filter.}

\item{client}{A \code{vmx_client}.}
}
\value{
A tibble.
}
\description{
List data versions
}
