% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{vmx_dataset_cancel}
\alias{vmx_dataset_cancel}
\title{Cancel a dataset's in-flight processing}
\usage{
vmx_dataset_cancel(dataset, client = vmx_client())
}
\arguments{
\item{dataset}{A dataset id or \code{vmx_dataset}.}

\item{client}{A \code{vmx_client}.}
}
\value{
A \code{vmx_dataset}.
}
\description{
Cancel a dataset's in-flight processing
}
