% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{vmx_datasets}
\alias{vmx_datasets}
\title{List datasets}
\usage{
vmx_datasets(study = NULL, treatment = NULL, client = vmx_client())
}
\arguments{
\item{study}{Optional study filter.}

\item{treatment}{Optional treatment filter.}

\item{client}{A \code{vmx_client}.}
}
\value{
A tibble containing all matching datasets.
}
\description{
List datasets
}
