% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_versions.R
\name{vmx_dosing}
\alias{vmx_dosing}
\title{Dosing events table}
\usage{
vmx_dosing(dv, client = vmx_client())
}
\arguments{
\item{dv}{A data-version id or \code{vmx_data_version}.}

\item{client}{A \code{vmx_client}.}
}
\value{
A tibble.
}
\description{
Dosing events table
}
