% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{vmx_fit_subject_estimates}
\alias{vmx_fit_subject_estimates}
\title{Subject-level parameter estimates}
\usage{
vmx_fit_subject_estimates(fit, client = vmx_client())
}
\arguments{
\item{fit}{A fit id or \code{vmx_model_fit}.}

\item{client}{A \code{vmx_client}.}
}
\value{
A tibble.
}
\description{
Subject-level parameter estimates
}
