% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{vmx_fit_vpc}
\alias{vmx_fit_vpc}
\title{Visual predictive check}
\usage{
vmx_fit_vpc(fit, client = vmx_client())
}
\arguments{
\item{fit}{A fit id or \code{vmx_model_fit}.}

\item{client}{A \code{vmx_client}.}
}
\value{
A ggplot-ready tibble.
}
\description{
Visual predictive check
}
