% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{vmx_health}
\alias{vmx_health}
\title{Connectivity probe}
\usage{
vmx_health(client = vmx_client())
}
\arguments{
\item{client}{A \code{vmx_client}.}
}
\value{
\code{TRUE} on success, otherwise a classed error.
}
\description{
Connectivity probe
}
