% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{vmx_model_build_export}
\alias{vmx_model_build_export}
\title{Markdown export of a build run}
\usage{
vmx_model_build_export(run, client = vmx_client())
}
\arguments{
\item{run}{A build-run id or object.}

\item{client}{A \code{vmx_client}.}
}
\description{
Markdown export of a build run
}
