% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{vmx_model_build_report}
\alias{vmx_model_build_report}
\title{Signed HTML report URL for a build run}
\usage{
vmx_model_build_report(run, client = vmx_client())
}
\arguments{
\item{run}{A build-run id or object.}

\item{client}{A \code{vmx_client}.}
}
\description{
Signed HTML report URL for a build run
}
