% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{vmx_model_build_runs}
\alias{vmx_model_build_runs}
\title{List model build runs}
\usage{
vmx_model_build_runs(
  data_version = NULL,
  study = NULL,
  treatment = NULL,
  status = NULL,
  client = vmx_client()
)
}
\arguments{
\item{data_version, study, treatment, status}{Optional filters.}

\item{client}{A \code{vmx_client}.}
}
\value{
A tibble containing all matching model-build runs.
}
\description{
List model build runs
}
