% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{vmx_model_catalog}
\alias{vmx_model_catalog}
\title{Model catalog}
\usage{
vmx_model_catalog(data_version = NULL, client = vmx_client())
}
\arguments{
\item{data_version}{Optional data-version filter.}

\item{client}{A \code{vmx_client}.}
}
\value{
A tibble.
}
\description{
Model catalog
}
