% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{vmx_model_fit}
\alias{vmx_model_fit}
\title{Fetch one model fit}
\usage{
vmx_model_fit(id, client = vmx_client())
}
\arguments{
\item{id}{A fit id or \code{vmx_model_fit}.}

\item{client}{A \code{vmx_client}.}
}
\value{
A \code{vmx_model_fit}.
}
\description{
Fetch one model fit
}
