% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{vmx_model_fits}
\alias{vmx_model_fits}
\title{List model fits}
\usage{
vmx_model_fits(
  run = NULL,
  data_version = NULL,
  model_type = NULL,
  marker_name = NULL,
  status = NULL,
  client = vmx_client()
)
}
\arguments{
\item{run, data_version, model_type, marker_name, status}{Optional filters.}

\item{client}{A \code{vmx_client}.}
}
\value{
A tibble.
}
\description{
List model fits
}
