% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nca.R
\name{vmx_nca_result}
\alias{vmx_nca_result}
\title{NCA result table (PK parameters)}
\usage{
vmx_nca_result(nca, client = vmx_client())
}
\arguments{
\item{nca}{An NCA id or \code{vmx_nca_analysis}.}

\item{client}{A \code{vmx_client}.}
}
\value{
A tibble.
}
\description{
NCA result table (PK parameters)
}
