% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_versions.R
\name{vmx_nlmixr_data}
\alias{vmx_nlmixr_data}
\title{NONMEM-layout data.frame for nlmixr2 / rxode2}
\usage{
vmx_nlmixr_data(dv, analyte = NULL, client = vmx_client())
}
\arguments{
\item{dv}{A data-version id or \code{vmx_data_version}.}

\item{analyte}{Analyte to assemble.}

\item{client}{A \code{vmx_client}.}
}
\value{
A data.frame ready for \code{nlmixr2()}.
}
\description{
NONMEM-layout data.frame for nlmixr2 / rxode2
}
