% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_versions.R
\name{vmx_pd}
\alias{vmx_pd}
\title{PD observations}
\usage{
vmx_pd(dv, marker = NULL, format = c("tidy", "nonmem"), client = vmx_client())
}
\arguments{
\item{dv}{A data-version id or \code{vmx_data_version}.}

\item{marker}{Optional PD marker filter.}

\item{format}{\code{"tidy"} or \code{"nonmem"}.}

\item{client}{A \code{vmx_client}.}
}
\value{
A tibble.
}
\description{
PD observations
}
