% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_versions.R
\name{vmx_pk}
\alias{vmx_pk}
\title{PK observations + events table}
\usage{
vmx_pk(dv, client = vmx_client())
}
\arguments{
\item{dv}{A data-version id or \code{vmx_data_version}.}

\item{client}{A \code{vmx_client}.}
}
\value{
A tibble.
}
\description{
PK observations + events table
}
