% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{vmx_sim_jobs}
\alias{vmx_sim_jobs}
\title{List simulation jobs for a model fit}
\usage{
vmx_sim_jobs(fit, client = vmx_client())
}
\arguments{
\item{fit}{A fit id or \code{vmx_model_fit}.}

\item{client}{A \code{vmx_client}.}
}
\value{
A tibble containing all simulation jobs for the fit.
}
\description{
List simulation jobs for a model fit
}
