% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_versions.R
\name{vmx_subjects}
\alias{vmx_subjects}
\title{Subjects table}
\usage{
vmx_subjects(dv, client = vmx_client())
}
\arguments{
\item{dv}{A data-version id or \code{vmx_data_version}.}

\item{client}{A \code{vmx_client}.}
}
\value{
A tibble, one row per subject.
}
\description{
Subjects table
}
