% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treatments.R
\name{vmx_treatments}
\alias{vmx_treatments}
\title{List treatments}
\usage{
vmx_treatments(status = NULL, client = vmx_client())
}
\arguments{
\item{status}{Optional status filter.}

\item{client}{A \code{vmx_client}.}
}
\value{
A tibble containing all matching treatments.
}
\description{
List treatments
}
