% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{vmx_upload_ignore}
\alias{vmx_upload_ignore}
\title{Ignore an upload}
\usage{
vmx_upload_ignore(upload, client = vmx_client())
}
\arguments{
\item{upload}{An upload id or object.}

\item{client}{A \code{vmx_client}.}
}
\description{
Ignore an upload
}
