% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{vmx_upload_unignore}
\alias{vmx_upload_unignore}
\title{Unignore an upload}
\usage{
vmx_upload_unignore(upload, client = vmx_client())
}
\arguments{
\item{upload}{An upload id or object.}

\item{client}{A \code{vmx_client}.}
}
\description{
Unignore an upload
}
